Rename composite module and fix broken tests
