+++ pyvista/__init__.py
@@ -17,7 +17,7 @@ from pyvista.grid import RectilinearGrid
 from pyvista.grid import UniformGrid
 from pyvista.geometric_objects import *
 from pyvista.parametric_geometry import *
-from pyvista.container import MultiBlock, CompositeFilters
+from pyvista.composite import MultiBlock, CompositeFilters
 from pyvista.qt_plotting import QtInteractor
 from pyvista.qt_plotting import BackgroundPlotter
 from pyvista.export import export_plotter_vtkjs, get_vtkjs_url
